% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-manual.R
\docType{data}
\name{defaults}
\alias{defaults}
\title{Defaults}
\format{
An object of class \code{character} of length 6.
}
\usage{
defaults
}
\description{
Default options for \verb{R CMD Rd2pdf}.
}
\seealso{
\code{\link{Rd2pdf}}
}
\keyword{datasets}
