% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-manual.R
\name{manual}
\alias{manual}
\title{Create R Manual}
\usage{
manual(path = ".", options = defaults, ...)
}
\arguments{
\item{path}{a list of file names specifying the R documentation sources
to use, by either giving the paths to the files, or the path to a
directory with the sources of a package.}

\item{options}{further options (as vector) to be passed to \verb{R CMD Rd2pdf}
(defaults to \code{defaults}).}

\item{...}{additional arguments passed to \code{system}.}
}
\value{
return value of \code{system}.
}
\description{
Calls \verb{R CMD Rd2pdf} with reasonable defaults.
}
\seealso{
\code{\link{Rd2pdf}}
}
